amazon nova 2/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-generativeai==0.8.3
pypdf==4.3.1
python-multipart==0.0.12
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.3.5
fastapi-users[sqlalchemy]==13.0.0
celery==5.4.0
redis==5.0.8
prometheus-client==0.21.0
alembic==1.13.2
slowapi==0.1.9